refactor(work): NewMessage 从 Session 按表拆分重构为 Work 业务领域独立服务
- 新建 WorkService/WorkHandler 接口和实现，按工作台业务域组织
- NewMessageRequest 移除 Title 字段，标题从 Content 截取
- 从 SessionService 接口和实现中移除 NewMessage 方法
- router.go 注册 WorkService 路由
